Divide SimpleProject in two Divided SimpleProject in two different classes: BaseSimpleProject and SimpleProject. The former will be in charge of managing implementations and dependencies, while the latter will only manage proxy instances creates/updates/admin changes
